Switched co-ordinate type, to line up with new absolute func

+++ 2017/03/02/day3_part2.go
@@ -18,7 +18,7 @@ const (
 )
 
 type spiralNodeCoords struct {
-	x, y int
+	x, y int64
 }
 
 type spiralNode struct {
